efl_debug (client): rewrite using efl_net.
Rewrite and greatly improve error checking, allow multiple commands to
be given in series, etc.

This also leverages on Efl.Io.Queue to buffer data for us, then all we
need to do is peek at its contents, when there is a full message we
use and then ask the queue to discard it.

The CLST processing was also rewritten to not allocate memory. Since
memory may be unaligned we still iterate over the buffer and memcpy()
to an integer to avoid alignment errors.