Makefile_Evas.am: (GL thread) Add a macro definition to classify building gl_common itself
If this macro is defined, every threaded GL API is builded as an exported symbol with function body.
But if this macro is not defined (outside of gl_common), every threaded GL API is builded as
a function pointer that links to exported gl_common symbols. (see LINK2GENERIC macros.)